+++ src/Views/Traits/Helpers/ColumnHelpers.php
@@ -449,4 +449,9 @@ trait ColumnHelpers
     {
         return $this->customSlug !== null;
     }
+
+    public function getColumnLabelStatus(): bool
+    {
+        return $this->displayColumnLabel ?? true;
+    }
 }